% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classify_jplace.R
\name{bipartition}
\alias{bipartition}
\title{Branches of a bipartition.}
\usage{
bipartition(tips, phy)
}
\arguments{
\item{tips}{a numeric vector with indices of tree tips belonging to the bipartition.}

\item{phy}{an object of class \code{phylo}.}
}
\value{
A matrix of mode \code{logical} whose columne correspond to tree branches and its two columns indicate,
whether the branch belongs to the crown of the bipartition and the second whether it belongs to its stem.
}
\description{
Finds tree branches belonging to a bipartition defined by the specified collection of tips.
}
\details{
The function is intended primarily as an internal function of \code{epaclades} package,
so its input and output is tailored to the needs of \link{classify_jplace}.
}
