% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classify_jplace.R
\name{classify_jplace}
\alias{classify_jplace}
\title{Classification of \code{jplace} object.}
\usage{
classify_jplace(jp, clades, ancestral = TRUE)
}
\arguments{
\item{jp}{an object of class \code{jplace}.}

\item{clades}{a matrix or data frame defining classification of tips (1st column) to clades / bipartitions
(2nd column). The classification may not be comprehensive (not including all the tips), but the clades
must not be overlapping or mutually nested.}

\item{ancestral}{Logical, whether to classify ancestral branches; The default option is \code{TRUE},
but it is relevant only if the tree is rooted or the classification to clades is comprehensive
(otherwise it is switched to \code{FALSE} and issues a warning).}
}
\value{
A object of class \code{jplace} with additional components \code{clade} and \code{position}
in the \code{placements} data frame, which indicate classification of the branches.
}
\description{
Classifies tree edges with any phylogenetic placements to clades (or bipartitions).
}
