% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/root_jplace.R
\name{getbipartcodes}
\alias{getbipartcodes}
\title{Bipartition codes.}
\usage{
getbipartcodes(phy, ref)
}
\arguments{
\item{phy}{an object of class \code{phylo}.}

\item{ref}{a character vector with tip labels (may or may not be tip labels of \code{phy}).}
}
\value{
A character vector with unique codes of tree branches, which are based on bipartitions defined by the branches.
}
\description{
Assign codes to branches using the corresponding complementary sets of tips (bipartitions).
}
\details{
The function is intended primarily as an internal function of \code{epaclades} package,
so its input and output is tailored to the needs of \link{root_jplace}.
}
