% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/root_jplace.R
\name{root_jplace}
\alias{root_jplace}
\title{Rooting \code{jplace} object.}
\usage{
root_jplace(jp, outgroup = "outgroup")
}
\arguments{
\item{jp}{an object of class \code{jplace}.}

\item{outgroup}{a character vector listing outgroup sequences or their unique identifier(s),
e.g., a character string "outgroup", which is the default.}
}
\value{
A modified \code{jplace} object, i.e., its \code{tree} is rooted, \code{outgroup} component added to it,
\code{placements$edge_num} is modified and if any placements were to the root edge,
their rows in \code{placements} are modified too.
}
\description{
Roots the tree & adjusts edge labelling accordingly.
}
